emitter visuals update (#13382)
